Cleanup and minor improvements.
-Renaming Files
-Fixing few issues with textures not showing up
-Controller Support

+++ C/SpaceShooter/game/source/entities/player.h
@@ -2,7 +2,7 @@
 #include "../../engine/engine.h"
 
 typedef struct Player{
-	Vec2i position;
+	Vec2f position;
 	SDL_Texture* texture;
 	bool isDead;
 	int score;
